requirements.txt
python-dotenv
mariadb
pandas
matplotlib
seaborn
scipy
scikit-learn

